Extract NewPlay item
